We don't really need FiatShamirTree module
