+++ hschain-utxo-lang/hschain-utxo-lang.cabal
@@ -69,7 +69,6 @@ library
       Hschain.Utxo.Lang.Sigma.DLog
       Hschain.Utxo.Lang.Sigma.EllipticCurve
       Hschain.Utxo.Lang.Sigma.Interpreter
-      Hschain.Utxo.Lang.Sigma.FiatShamirTree
       Hschain.Utxo.Lang.Sigma.MultiSig
       Hschain.Utxo.Lang.Sigma.Protocol
       Hschain.Utxo.Lang.Sigma.Types
